// seanime-web-legacy/src/api/generated/endpoints.ts
export const API_ENDPOINTS = {
// ...
    ANIME_COLLECTION: {
        /**
         *  @description
         *  Route returns the main local anime collection.
         *  This creates a new LibraryCollection struct and returns it.
         *  This is used to get the main anime collection of the user.
         *  It uses the cached Anilist anime collection for the GET method.
         *  It refreshes the AniList anime collection if the POST method is used.
         */
        GetLibraryCollection: {
            key: "ANIME-COLLECTION-get-library-collection",
            methods: ["GET", "POST"],
            endpoint: "/api/v1/library/collection",
        },
        /**
         *  @description
         *  Route returns anime collection schedule
         *  This is used by the "Schedule" page to display the anime schedule.
         */
        GetAnimeCollectionSchedule: {
            key: "ANIME-COLLECTION-get-anime-collection-schedule",
            methods: ["GET"],
            endpoint: "/api/v1/library/schedule",
        },
        /**
         *  @description
         *  Route adds the given media to the user's AniList planning collections
         *  Since media not found in the user's AniList collection are not displayed in the library, this route is used to add them.
         *  The response is ignored in the frontend, the client should just refetch the entire library collection.
         */
        AddUnknownMedia: {
            key: "ANIME-COLLECTION-add-unknown-media",
            methods: ["POST"],
            endpoint: "/api/v1/library/unknown-media",
        },
    },
// ...
    AUTH: {
        /**
         *  @description
         *  Route logs in the user by saving the JWT token in the database.
         *  This is called when the JWT token is obtained from AniList after logging in with redirection on the client.
         *  It also fetches the Viewer data from AniList and saves it in the database.
         *  It creates a new handlers.Status and refreshes App modules.
         */
        Login: {
            key: "AUTH-login",
            methods: ["POST"],
            endpoint: "/api/v1/auth/login",
        },
        /**
         *  @description
         *  Route logs out the user by removing JWT token from the database.
         *  It removes JWT token and Viewer data from the database.
         *  It creates a new handlers.Status and refreshes App modules.
         */
        Logout: {
            key: "AUTH-logout",
            methods: ["POST"],
            endpoint: "/api/v1/auth/logout",
        },
    },
// ...
} satisfies ApiEndpoints